fix to subsampling half of subjects
In editing a script that was derived from one of these, I realized the 'fidelity' analysis that quantifies fidelity over multiple timepoints during a delay actually samples only half of subjects (because we're each trial contributes two timepoints per delay). I've fixed this error, which only applied to the fidelity analysis per delay, and results are qualitatively the same - in fact stronger.